pyproject.toml
[project]
name = "grove"
version = "1.0.0-dev"
dependencies = [
    "django~=5.0",
    "plastron-utils~=4.0",
    "rdflib~=7.0",
]

[project.optional-dependencies]
test = [
    "pytest~=8.1",
    "pytest-cov~=4.1",
    "pytest-datadir~=1.5",
    "pytest-django~=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "grove.settings"

[tool.coverage.run]
# these were generated by Django
omit = [
    "asgi.py",
    "manage.py",
    "wsgi.py",
]
